Update Get-MMSSessionContent.ps1
Funny, I was looking at doing something like this last week and saw your twitter post about it so took a look. The proposed changes just simplify it a little and also allow to create a "Session Info.txt" file with some details about each session inside the folders - which I thought could be handy. I think it could be even better by removing the cabana type sessions...but didn't look to see if all conferences use the same format or not so figured having them wouldn't be that big of a deal for now. Whatcha think?